requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
psutil==5.9.7
GPUtil==1.4.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.0
spacy==3.7.2
networkx==3.2.1
pydantic==2.5.3
python-dotenv==1.0.0
